refactor(advisor): move KV advisor unit tests to a sibling _tests.rs
Split the inline `#[cfg(test)] mod tests` out of kv_cache_advisor.rs into kv_cache_advisor_tests.rs, wired with `#[path = "kv_cache_advisor_tests.rs"] mod tests;`. This follows the CLAUDE.md convention (unit tests in `_tests.rs` files beside the implementation) and keeps the implementation file under the 500-line limit. No code or test behavior changes; the tests still run as execution::kv_cache_advisor::tests::*.
